Simplify product API
